requirements.txt
django == 3.2.13
pillow
django-admin-interface
django-crispy-forms
django-rest-framework